+++ packages/embedded-client/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - [`OutputResult.error`](https://cerbos.github.io/cerbos-sdk-javascript/interfaces/_cerbos_core.OutputResult.html#error) property ([#1452](https://github.com/cerbos/cerbos-sdk-javascript/pull/1452))
 
+### Changed
+
+- Bump dependency on [@cerbos/api] to 0.9.0 ([#1460](https://github.com/cerbos/cerbos-sdk-javascript/pull/1460))
+
 ### Removed
 
 - Support for Node.js 20, which is now [end-of-life][nodejs-eol] ([#1454](https://github.com/cerbos/cerbos-sdk-javascript/pull/1454))