+++ packages/embedded-client/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- Bump dependency on [@bufbuild/protobuf] to 2.12.1 ([#1478](https://github.com/cerbos/cerbos-sdk-javascript/pull/1478))
+
 - Bump dependency on [@cerbos/api] to 0.9.1 ([#1462](https://github.com/cerbos/cerbos-sdk-javascript/pull/1462))
 
 ### Removed